Moved CheckReminder to a module folder, moved CheckReminderData class to the globals
