+++ lib/src/bot.dart
@@ -18,7 +18,7 @@ import 'package:weather/src/modules/conversator_manager.dart';
 import 'package:weather/src/modules/general_manager.dart';
 import 'package:weather/src/modules/accordion_poll/accordion_poll_manager.dart';
 import 'package:weather/src/modules/command_statistics_manager.dart';
-import 'package:weather/src/modules/check_reminder_manager.dart';
+import 'package:weather/src/modules/check_reminder/check_reminder_manager.dart';
 
 class Bot {
   final Config _config;
