+++ lib/src/globals/check_reminder_data.dart
@@ -1,0 +1,9 @@
+class CheckReminderData {
+  final int id;
+  final String chatId;
+  final String userId;
+  final String message;
+  final DateTime timestamp;
+
+  CheckReminderData({required this.id, required this.chatId, required this.userId, required this.message, required this.timestamp});
+}
